fixed: do not wrongly hide some ALSA HDMI devices
Currently the AESinkALSA code takes empty but existing ELD data to
always mean that the HDMI port is unconnected and hides it.

However, many HDA HDMI output devices do not properly pass ELD data,
leaving the reported ELD empty.

Therefore, only hide unconnected ports for HDA Intel and HDA NVidia,
since ELD seems to be reliable with those cards/drivers (i.e. if it
exists, it is populated, and if it is not supported, it does not exist).