CacheMemBuffer Feature Addition:
Added an advanced settings variable "cachemembufferthrottlemultiplier" which allows the user to fill the cachemembuffer at a multiplier of the average bitrate of the video. This will lead to faster buffer fills (i.e. if user uses 3 as the multiple the buffer will be filled at 3X average bitrate until the buffer is filled and then will refill the buffer as it is being used at up to 3X average bitrate)

NOTE: The variable "cachemembufferthrottlemultiplier" will need to be added to the advanced setting wiki.  It should be placed in network section under cachemembuffer like so:

<network>
<cachemembufferthrottlemultiplier>1</cachemembufferthrottlemultiplier>  <!-- Adjusts the filling throttle of cachemembuffer by X times average bitrate -->
</network>