fixed: too many ALSA audio devices enumerated on Ubuntu
Ubuntu has patched their alsa-lib so that "defaults.namehint.extended"
defaults to "on" instead of upstream "off", causing lots of unwanted
extra devices (many of which are not actually routed properly, like the
"dmix:$foo" ones which may actually direct to a completely different
device) to be found by the enumeration process.

Skip such devices.