Fix XBMC not fully exiting when Alt-F4 is pressed during video playback.
Fixes #13177.

The following was happening:

- Pressing Alt-F4 triggers WM_CLOSE and then WM_DESTROY.  g_hWnd is no longer
valid after WM_DESTROY, but is still set.
- AESinkDirectSound changes format, due to the video being stopped.
CAESinkDirectSound::Initialize's m_pDSound->SetCooperativeLevel call silently
fails: it returns success but doesn't actually do anything.  This is triggered
by us passing in the now-invalid g_hWnd. (Not returning an error may be a
DirectSound bug.)
- CAESinkDirectSound::CheckPlayStatus attempts to start playback, and fails
with DSERR_PRIOLEVELNEEDED.  This is a silent failure; the result wasn't being
checked or logged.
- CAESinkDirectSound::AddPackets gets stuck in the while(GetSpace() < total)
loop, because the buffer isn't actually playing.
- CAEFactory::Shutdown never returns, because the sink never shuts down.

To fix this, ensure that g_hWnd is cleared on WM_DESTROY, so we don't pass in
bogus values to SetCooperativeLevel.  This can't be done in
CWinSystemWin32::DestroyWindowSystem, since it happens too late.  Do it in
the event handler, so it's cleared as soon as it's invalidated.