[AE][WASAPI] fixes #1287 crash in WASAPI sink after changes made to unify return values of GetCacheTime.
Due to the event-driven mode of WASAPI used, where twin buffers are operated ping-pong style and upon notification of a buffer-free event an entire buffer must be written, the cache is always 50% or 100% full.

This causes issues with water-levels in the players and streams. Two methods have been tested to track the gradient cache time:
1) use of the system timer to decrement the cached duration between 100% and just under 50% resetting it each time a new buffer is written, and
2) tracking the actual audio driver's buffer pointer using non-standard system calls.

Neither solve the issue of water-levels > 50% and add significant overhead. Ultimately the returned value seems to have no impact on A/V sync and the correct buffering is achieved via AddPackets() and the buffer-empty events used in WASAPI.

This makes this sink unique from the others including DirectSound where it is possible to adhere to the function definitions. This has been clearly commented in the code.