[pvr] revert some of the changes to old HTSP client and make PVR manager use IDisplayTime
The changes seemingly made no sense, and this simplifies code.

Note, there is a bug here. GetStartTIme() originally meant: how much time has elapses from the time we started playing this channel and the start of the current playing program. MythTv file still uses this definition but PVRManager does not. So if PVRManager ever uses MythTv as a backing file, it will be wrong.