Fix possible NULL pointer dereference in CGUIDialogPVRGuideInfo::ActionStartTimer()
I'm not sure if we should return early if tag is NULL, or just avoid
setting the lines on the dialog.  I chose to avoid setting the lines as
tag is passed in as a parameter so it may be possible for it to be
optional.

Guidance on this change is apprecaited.