d-r: enable dirty-regions by default
This enables the all-or-nothing mode, which means that we repaint the entire
screen if there's a single dirty region.

Also, nofliptimeout=0 is set to ensure that we never flip without rendering.

With these combined, this should be universally safe (and within spec) for all
gfx cards/drivers.

The main possible side-effect is the slow-motion effect brought about on slower
cpus that can't keep up at times and produce wild timestamps. With our improved
decoding and caching, this is now quite rare.