build: bootstrap overhaul
Put an end to the constant autoreconfs when switching branches. Use make to
determine when bootstrapping for xbmc and subprojects needs to happen
(hint: rarely).

bootstrapping can now be done in parallel, but I didn't obseve much (any?)
speed increase.

Also allows for easier skipping of reconfigures for minor/cosmetic configure.in
changes.

This should be transparent to users, other than the fact that it's much faster.

Same build procedure as always: ./bootstrap && ./configure && make