[AE][SoftAE] Fixed segfaults & improved WAV handling
Previously when a sound was played, if the format did not match the sink format
it would be re-loaded from the disk and re-sampled each time the sink re-opened.
This behaviour has been changed so that the samples are cached along with the
original format information. This will reduce the re-init latency.

A new method has been added to check if the current output WAV format is
compatible with the required output format, so that re-init is not performed
when it is not required. This fixes the soft suspend issue with SoftAE in
CSoftAE::MixSounds which caused stack smashing.

This also fixes a long standing bug in CSoftAE::MixSounds when multiple sounds
are played at once that was also causing stack smashing.