egl: rewrite
As promised with f6ecaea0, here is a complete rewrite of EGL. At the moment
this affects only android/amlogic/rpi.

Our previous system of inheritance was wonky, and was one hack after another.
This provides a single coherent implementation of the window life-cycle. It
also begins the process of decoupling rendering from windowing (at least for
EGL)

Each client is expected to fill in its windowing functionality outside of egl.

In theory, adding a new back-end requires only describing how to create/destroy
native structures and enumerate resolutions. If more functionality is needed, it
should be added to the EGLNativeType interface rather than hacking up the EGL
implementation itself.

Also, the back-ends are detected and tested at runtime rather than build-
time, meaning that it is not necessary to specify a back-end. It also
means that a single binary would be capable of running on multiple EGL
back-ends (for example, fbdev and X) when they're added.

Each back-end is required to provide a test-function at runtime to
determine whether it should be considered for use.

TODO: some introspect for enumerating possible back-ends before window-creation
TODO: fall-back if the first back-end fails
TODO: add rpi back-end