egl: Create a way to bypass render completely
If render is disabled, we essentially go headless. This is necessary on android
because when our window is destroyed, all graphics operations become errors.

This also has the effect of dropping cpu usage down to ~0% while we're running
in the background, so it could be used on other platforms when minimized.