Revert "Check return of stat64 from XFileUtils.cpp"
This reverts commit d31305cab62e67399c65a0bddfd123f3947ec1f8.

Returning FALSE from FindNextFile() means that there are no more files
in the directory. Therefore a failing stat64() call should not cause
FALSE to be returned. It is also not a fatal error, it will just mean we
do not get much actual information about the file.

stat64() can fail e.g. on broken symlinks, so this commit caused files
in directories that contain broken symlinks to be only partially shown.