Picture database built on dynamic database abstraction
Backend is a denormalized database with dynamic normalization for the purpose of achieving optimal read queries. Objects are stored in a central table with key/value pairs. When 1:1, 1:N and N:N relationships are declared, data is drawn from the key/value pairs to fill the newly created tables. Data is kept in sync across adds and deletes, as well as when new relations are added/dropped.