[rbp] Speed, quality and reliability fixes for images.
fixes a couple of bugs in the jpeg header parsing, which would sometimes get incorrect dimensions or no dimension.
It avoids the resize to a multiple of 16 of decoded jpegs which causes blurring and subtle aspect ratio erros.
It avoids a pitch mangling alloc and copy on the jpeg encode path.
The jpeg decoder now handles the port settings changes event properly which can fix an OMX error case.
An updated start.elf is required which supports the non-multiple of 16 pitches in openmax.