fixed: ssa subtitles where packets overlap in time
Sometimes a ssa packet would show up that had a start
time before the end time of previous packet, but a
end time before previous packet's end time.

To solve this with auto replacing packet, we need
to extend the end time of the new packet to include
the old packet.

This fixes ticket: #13302