[droid] Hack Hack. Stop AE from chewing up the cpu
I really don't like this, but it's the safest thing we could come up with for
Frodo. On my hardware, cpu usage drops from 45% to 15% when idling when this
change is made. Presumably we should be sleeping somewhere in the Android sink
that we are not doing currently.

For Android, we should be suspending, but it causes a race that is too complex
to address at this point in the release cycle. When we do suspend, cpu usage
drops further to 8%.

This needs to be DROPPED post-Frodo, and completely re-evaluated.