addons: fix possible deadlock which causes exit to fail
If an PVR addon has ever lost its connection, there's a race condition that can
cause the thread to never die, and we deadlock when trying to exit XBMC. This
should probably be fixed somewhere in threading so that the GetPriority() and
SetPriority() calls are safer, but in this case we'll just make sure that the
thread is running when we call it.

The underlying problem is that SetPriority() waits on the thread-start event,
which is in a weird state if the thread is set to auto-delete and it has a very
short life-span. In that case, the event never shows as signaled in
SetPriority(). As a result the calling thread waits forever for the ghost
thread to start, and hangs there forever.