Improve cache handling for streams
The old code was ratelimiting and keeping the same cachesize for all
files. The new code is NOT ratelimiting at all, I want to try this
before we change it, I have not seen any problem removing the limit
locally, but I guess it's there for a reason.

The cache limit is now set by the size of the file, this means that we
can create the cache in the correctsize instead of like before waiting
for the bitrate.

The default values are that we cache 5% of the filesize or at maximum
<smartcacheupperlimit> from advancedsettings.

Related to #415