Fix audioqueue starvation
We used a WAAAAY to small buffer for the videoqueue, this lead to
strangely enough problems with the audioqueue, I think it's because
when you have a very highres / bandwidth intensive video it will take
priority in the queues. Anyway this was legacy code that we are better
of without.

Fixes #460