Update xbmc/input/KeyboardStat.cpp
Format of keyname must NOT be hex. The usage in keyboard.xml in the keymap directory is also NOT hex. This is giving very much trouble for people who want to use keyboard.xml. I spend more than an hour on this to figer it out as a non XBMC developer.