[droid] attach/detach the jvm using thread storage
Threads are automatically attached when issuing a xbmc_jnienv(), and detach
automatically at destruction.

This eliminates manual intervention, and likely will save us from some bugs in
the future. The old code was very difficult to maintain since every function
had to assume that it was not attached and had to detach when finished.

This is the method recommended in the Android docs.