ffmpeg forward compatibility fix for matroska embedded subtitles
Commit 2626cc4580bfd560c6983338d77b2c11c16af94f in the ffmpeg repo
changed the codec id of matroska embedded S_TEXT/UTF-8 subtitles
from CODEC_ID_TEXT to AV_CODEC_ID_SUBRIP. That commit is part of
ffmpeg 1.0 and later.

Add the new codec id to provide forward compatibility for when
the embedded ffmpeg code is updated and/or when building with
external ffmpeg libraries.