Fix minor CenterRating issues
Some of these CenterRating had some "rogue" pixels and on left right of
some (not all) of the stars, ended up redoing the lot just copy and paste the stars into place
because of these pixels and also reduced the size of the png as well as result.

The change only affected some of the graphics and hard to spot but it is/was there.

Test Textures.xbt is updated grab it from usual link