rendermanager: remove inheritance from CBaseTexture.
Add a "RenderObject" member that represents an abstract gpu texture handle.
It is defined as a void*, and should be cast by each implementation to whatever
it represents. For example, gl/gles both define it as a GLuint (unsigned int).

It should NOT have a hard-coded build-time definition.

Also, add some helper functions for getting at the texture easier.