CGUIWindowSettingsCategory: only handle delayed settings during GUI_MSG_FOCUSED if the focused setting is not the delayed setting
When using a mouse to control XBMC changing the value of a spinner with a left
click results in a GUI_MSG_FOCUSED message. If that setting is delayed a left
click lead to the early evaluation of the delayed setting even though it's
still being changed (which should just result in a reset of the delay timer).
This lead to crashes if a setting changed callback executed a skin reload.