CGUIWindowSettingsCategory: update delayed setting changes on focus change asynchronously
When using the mouse to focus another control after having changed the value
of a delayed setting that reloads the skin, a SIGSEGV occured because the
CGUIControl event handler tried to execute code on an already destroyed object.
Handling the update of the delayed setting change asynchronously allows the
active event handler to finish working before handling the previously triggered
GUI_MSG_UPDATE_ITEM which may lead to the destruction of the setting controls.