[rbp] Disable unimplemented settings options
I've been through the settings carefully and have made any settings that are not implemented on Pi be not-visible.

If and when support is added for probing capabilities of the media player then the hiding of unsupported options can be removed.

Thanks to montellese for making this configuration so simple.