[rbp/omxplayer] Ignore video stream in audio filetypes
Since the ffmpeg 1.2 merge, we fail to play audio file with artwork.
These show up as having a mjpeg video stream, with just one frame in.
We get stuck in buffering and don't play the audio.

As people expect the paplayer behaviour, we ignore the video stream for
audio filetypes, so the audio can play with visualision.