pvr: Pass selected item to menu hooks
Currently it is not possible to retrieve any information about the
selected item in the addon's menu hook callback. Therefore menu
hooks can only trigger actions per category, but not per item.

This patch removes this limitation by extending CallMenuHook with
an item parameter representing the type (PVR_MENUHOOK_CAT) and the
data of the currently selected item. The type PVR_MENUHOOK_UNKNOWN
indicates that no selection is available (as for settings hooks).