pvr: win32: Make sure that the addon libs define TARGET_WINDOWS
Recently, _WIN32 has been replaced by TARGET_WINDOWS.
Since this is a project specific define and not a system variable,
we have to make sure that the PVR addon libs define this in their
project configuration. Otherwise the libs won't export any symbols
and the PVR addons cannot be loaded.