Don't add xbmc.a to DIRECTORY_ARCHIVES
It gets added to MAINOBJS anyways, and adding it to DIRECTORY_ARCHIVES
would mean that it also gets added to the tests, which would result
in a double-definition of main () (one from xbmc and one from gtest).

Instead, add it to xbmc.bin directly, as Darwin based platforms require
such static linking.