Built in function PlayMedia does not play music folders
Following specification, PlayMedia should be able of playing content of
folders, either with content of video or music.

An issue has been found by which passing a music folder to this
function won't play its content. However, if the folder contains video
files it works.

This function internally enumerates files by extension, and when a
folder is received, only video extensions are searched. This commit add
music extensions too to the search, solving the issue.