Listen for a new wl_keyboard from wl_seat_listener and create an XKBKeymap implementing wayland::Keymap. This class keeps track of the keyboard state and MUST be updated every time a modifier or a key is pressed (eg, every time we get an event from the protocol).
KeyboardProcessor implements the logic of converting XKB keysyms to
XBMC keysyms. It also changes the focus state of the window and
delivers key events to the XBMC event subsystem.

TimeoutManager and WaylandEventLoop are just a simple abstraction for
adding timeouts to the poll () / wl_display_dispatch () mainloop, in
place so that we can implement key repeat