music: fix crash when calling CTagLoaderTagLib::Load() through IMusicInfoTagLoader::Load() (introduced by d0a5a97e0c6c7831db382dfdc3e8650ff500acb9)
The class definition of CTagLoaderTagLib was completely screwed up. It is used
as if it was derived from IMusicInfoTagLoader but it actually wasn't.
Furthermore the signature of the CTagLoaderTagLib::Load() method was not
completely identical to the pure virtual method definition in
IMusicInfoTagLoader (std::string instead of CStdString) so it's a miracle this
didn't blow into our faces earlier.