Read or dispatch events in a separate thread.
MessagePump() should not be a blocking operation

Because wl_display_read_events was only introduced in wayland 1.2, we need
to support two methods of making MessagePump() nonblocking.

version_11::EventQueueStrategy dispatches the whole event queue and gathers
the resulting XBMC_Events or other messages wrapped up into Action functors
in its reader thread. It then dispatches these to XBMC in the main thread.

version_12::EventQueueStrategy simply reads the wayland connection fd and
then flushes the output buffer and dispatches any read events in the
main thread.

In either version, the wayland listeners may be running in a separate thread.

As such, there is no guaruntee that it is safe to mutate the state of
objects in the xbmc::wayland namespace, at least until
version_11::EventQueueStrategy can be dropped.
XBMCConnection::WaitForSynchronize() should also only be called from
the main thread and not the dispatch thread as it is responsible for
setting synchronized == true