Add wayland object listeners.
We will need direct access to some wayland objects in testing,
some of which are only available to the client itself. Add some
hooks to the client classes to advertise via a singleton via
any other interested code that some wayland object became available.

Test that we received the xbmc_wayland object after connection and
sync.