Fix updater test behavior with Ruby 1.9.x
The behavior of FileUtils.cp w.r.t. preserving the source file
mode changed between Ruby 1.8 and Ruby 1.9, in particular the execute
permission was not preserved for copied binaries so the test script
failed to launch them.

Set the :preserve attribute so that the source file's mode is
used.

MD-19678 #time 1h